feat: the get endpoints now return a better response

+++ src/database/prisma/repositories/prisma-authors.repository.ts
@@ -1,12 +1,12 @@
-import { Author } from '@/entities/author';
-import { AuthorsRepository } from '@/modules/authors/authors-repository';
-import { Injectable } from '@nestjs/common';
+import { Author } from '@/database/entities/author';
+import { AuthorsRepository } from '@/modules/authors/authors.repository';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaAuthorMapper } from '../mappers/prisma-author.mapper';
 import { PrismaService } from '../prisma.service';
 
 @Injectable()
 export class PrismaAuthorsRepository implements AuthorsRepository {
-  constructor(private prisma: PrismaService) {}
+  constructor(private readonly prisma: PrismaService) {}
 
   async createAuthor(authorData: Author): Promise<void> {
     const raw = PrismaAuthorMapper.toPrisma(authorData);
@@ -17,20 +17,44 @@ export class PrismaAuthorsRepository implements AuthorsRepository {
   async getAuthorById(authorId: number): Promise<Author> {
     const author = await this.prisma.author.findUnique({
       where: { id: authorId },
+      include: {
+        books: true,
+      },
     });
 
     if (!author) {
-      return null;
+      throw new NotFoundException('Author not found');
     }
 
     return PrismaAuthorMapper.toEntity(author);
   }
 
-  async updateAuthor(authorId: number, author: Author): Promise<void> {
-    throw new Error('Method not implemented.');
+  async updateAuthor(authorId: number, authorData: Author): Promise<void> {
+    const book = await this.prisma.author.findUnique({
+      where: { id: authorId },
+    });
+
+    if (!book) {
+      throw new NotFoundException('Author not found');
+    }
+
+    const raw = PrismaAuthorMapper.toPrisma(authorData);
+
+    await this.prisma.author.update({
+      where: { id: authorId },
+      data: raw,
+    });
   }
 
   async deleteAuthor(authorId: number): Promise<void> {
+    const book = await this.prisma.author.findUnique({
+      where: { id: authorId },
+    });
+
+    if (!book) {
+      throw new NotFoundException('Author not found');
+    }
+
     await this.prisma.author.delete({ where: { id: authorId } });
   }
 }
